backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
attrs==25.4.0
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
distlib==0.4.0
distro==1.9.0
dj-database-url==3.1.0
Django==5.2.6
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
dlib==20.0.0
dnspython==2.8.0
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.19.1
fonttools==4.61.1
fsspec==2025.12.0
future==1.0.0
groq==0.34.1
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huepy==1.2.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
instabot==0.117.0
Jinja2==3.1.6
jiter==0.12.0
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mock==5.2.0
moviepy==2.2.1
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
openai==2.8.1
opencv-python==4.12.0.88
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
proglog==0.1.12
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.15.0
PyMySQL==1.1.2
pyparsing==3.2.5
PyPDF2==3.0.1
PySocks==1.7.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
responses==0.25.8
schedule==1.2.2
selenium==4.38.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
sympy==1.14.0
torch==2.9.1
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
websocket-client==1.9.0
wheel==0.45.1
wsproto==1.3.2
